+++ packages/extension-base/src/background/RequestBytesSign.ts
@@ -6,7 +6,7 @@ import type { SignerPayloadRaw } from '@polkadot/types/types';
 import type { HexString } from '@polkadot/util/types';
 import type { RequestSign } from './types';
 
-import { wrapBytes } from '@reef-defi/extension-dapp/wrapBytes';
+import { wrapBytes } from '@dust-defi/extension-dapp/wrapBytes';
 
 import { TypeRegistry } from '@polkadot/types';
 import { u8aToHex } from '@polkadot/util';