+++ packages/extension-ui/src/Popup/Derive/SelectParent.tsx
@@ -1,7 +1,7 @@
 // Copyright 2019-2021 @polkadot/extension-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { canDerive } from '@reef-defi/extension-base/utils';
+import { canDerive } from '@dust-defi/extension-base/utils';
 import React, { useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react';
 
 import { AccountContext, ActionContext, Address, ButtonArea, InputWithLabel, Label, NextStepButton, VerticalSpace, Warning } from '../../components';