+++ packages/extension-ui/src/Popup/Signing/Qr.tsx
@@ -4,7 +4,7 @@
 import type { ExtrinsicPayload } from '@polkadot/types/interfaces';
 import type { HexString } from '@polkadot/util/types';
 
-import { wrapBytes } from '@reef-defi/extension-dapp/wrapBytes';
+import { wrapBytes } from '@dust-defi/extension-dapp/wrapBytes';
 import React, { useCallback, useMemo, useState } from 'react';
 import styled from 'styled-components';