+++ packages/extension-ui/src/Popup/index.tsx
@@ -1,11 +1,11 @@
 // Copyright 2019-2021 @polkadot/extension-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { AccountJson, AccountsContext, AuthorizeRequest, MetadataRequest, SigningRequest } from '@reef-defi/extension-base/background/types';
+import type { AccountJson, AccountsContext, AuthorizeRequest, MetadataRequest, SigningRequest } from '@dust-defi/extension-base/background/types';
 import type { SettingsStruct } from '@polkadot/ui-settings/types';
 
-import { PHISHING_PAGE_REDIRECT } from '@reef-defi/extension-base/defaults';
-import { canDerive } from '@reef-defi/extension-base/utils';
+import { PHISHING_PAGE_REDIRECT } from '@dust-defi/extension-base/defaults';
+import { canDerive } from '@dust-defi/extension-base/utils';
 import React, { useCallback, useEffect, useState } from 'react';
 import { Route, Switch } from 'react-router';