+++ packages/extension-ui/src/components/Address.test.tsx
@@ -3,7 +3,7 @@
 
 import '../../../../__mocks__/chrome';
 
-import type { AccountJson } from '@reef-defi/extension-base/background/types';
+import type { AccountJson } from '@dust-defi/extension-base/background/types';
 import type { ReactWrapper } from 'enzyme';
 import type { IconTheme } from '@polkadot/react-identicon/types';
 import type { Props as AddressComponentProps } from './Address';