+++ packages/extension-ui/src/messaging.test.ts
@@ -3,7 +3,7 @@
 
 import '../../../__mocks__/chrome';
 
-import chrome from '@reef-defi/extension-inject/chrome';
+import chrome from '@dust-defi/extension-inject/chrome';
 import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 import { configure } from 'enzyme';